requirement.txt
appdirs==1.4.4
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
darknet==0.3
distlib==0.3.1
docopt==0.6.2
filelock==3.0.12
Flask==1.1.2
Flask-Login==0.5.0
Flask-SQLAlchemy==2.5.1
greenlet==1.0.0
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.3
MarkupSafe==1.1.1
mysql-connector-python==8.0.23
mysqlclient==2.0.3
numpy==1.20.2
opencv-python==4.5.1.48
pipreqs==0.4.10
protobuf==3.15.8
PyMySQL==1.0.2
requests==2.25.1
six==1.15.0
SQLAlchemy==1.4.5
urllib3==1.26.4
virtualenv==20.4.3
Werkzeug==1.0.1
yarg==0.1.9
